requirements.txt
torch>=1.4
transformers
pinecone-client
sentence-transformers
scipy
ftfy
accelerate
pyngrok==4.1.1
flask_ngrok